Module Libelectrod.Location

type t

A location in a file represents an interval from a start position to an end one.

val from_positions : Stdlib.Lexing.position -> Stdlib.Lexing.position -> t

from_positions begp endp takes Lexing.positions begp and endp position returns a location out of them. Requires: begp < endp.

val begl : t -> int

val begc : t -> int

val endl : t -> int

val endc : t -> int

val to_ints : t -> (int * int) * (int * int)

val span : (t * t) -> t

Merge two positions

val dummy : t

Dummy position

type 'a located = {

data : 'a;

loc : t;

}

val make_located : 'a -> t -> 'a located

val pp_located : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a located -> unit

include Intf.Print.S with type t := t

type t

val pp : t Libelectrod.Fmtc.t

val to_string : t -> string