Module Libelectrod__Domain

The domain represents the set of relation declarations.

type t

Virtually: a map between relation names and their definition as sets of tuples.

val empty : t

Constructor.

val add : Libelectrod.Name.t -> Libelectrod.Relation.t -> t -> t

Adds an asociation to the domain. The name must not be in the domain already.

val mem : Libelectrod.Name.t -> t -> bool

Checks whether a name is already bound in the map.

Accessors

val get_exn : Libelectrod.Name.t -> t -> Libelectrod.Relation.t

May raise Not_found.

val get : Libelectrod.Name.t -> t -> Libelectrod.Relation.t option

May rather return None.

val univ_atoms : t -> Libelectrod.Tuple_set.t

Returns the set of atoms in univ, represented as a Tuple_set.t

val to_list : t -> (Libelectrod.Name.t * Libelectrod.Relation.t) list

Returns the map as an association list

val of_list : (Libelectrod.Name.t * Libelectrod.Relation.t) list -> t

val equal : t -> t -> bool

val must : Libelectrod.Name.t -> t -> Libelectrod.Tuple_set.t

Returns the "may" and "must" tuple sets associated to a relation name.

val may : Libelectrod.Name.t -> t -> Libelectrod.Tuple_set.t

val sup : Libelectrod.Name.t -> t -> Libelectrod.Tuple_set.t

val musts : ?⁠with_univ_and_ident:bool -> t -> (Libelectrod.Name.t, Libelectrod.Tuple_set.t) Containers.List.Assoc.t

Returns the association list between relation names and their "must" set.

val arities : t -> (Libelectrod.Name.t, int) Containers.List.Assoc.t

Returns the association list between relation names and their arity.

val rename : (Libelectrod.Atom.t, Libelectrod.Atom.t) Containers.List.Assoc.t -> (Libelectrod.Name.t, Libelectrod.Name.t) Containers.List.Assoc.t -> t -> t

Returns a new domain where atoms and relations have been renamed.

val update_domain_with_instance : t -> Libelectrod.Instance.t -> t

For every entry in inst, update_domain_with_instance dom inst replaces the corresponding relation in dom with the exact scope given by inst.

include Libelectrod.Intf.Print.S with type t := t

type t

val pp : t Libelectrod.Fmtc.t

val to_string : t -> string