Module Libelectrod__Elo_to_model1.Make

Parameters

Ltl : Libelectrod.Solver.LTL

ConvertFormulas : Libelectrod.Elo_to_ltl_intf.S with type ltl = Ltl.t and type atomic = Ltl.Atomic.t

Model : Libelectrod.Solver.MODEL with type ltl = ConvertFormulas.ltl and type atomic = ConvertFormulas.atomic

Signature

type atomic = Ltl.Atomic.t

val single_state_sym_to_ltl : int -> Libelectrod.Elo.t -> Libelectrod.Symmetry.t -> Ltl.t

Computes the LTL (the only temporal connective is "next") formula encoding the symmetry in a single state only (int the initial state if symmetry_offset = 0)

val temporal_sym_to_ltl : Libelectrod.Elo.t -> Libelectrod.Symmetry.t -> Ltl.t

Computes the full LTL formula encoding the symmetry in a temporal context

val syms_to_ltl : bool -> int -> Libelectrod.Elo.t -> Ltl.t Containers.List.t

val add_sym_comment_to_ltl_fml_list : 'a list -> (string * 'a) S.t

val split_invar_noninvar_fmls : Libelectrod.Elo.t -> Libelectrod.Elo.fml list -> Libelectrod.Elo.fml list * Libelectrod.Elo.fml list * Libelectrod.Elo.fml list * Libelectrod.Elo.fml list

val dualise_fmls : Libelectrod.Elo.fml list -> Libelectrod.Elo.fml

val run : (Libelectrod.Elo.t * bool * int) -> Model.t