Module Libelectrod__Instance

type t

Virtually: a map between relation names and sets of tuples.

val empty : t

Constructor.

val add : Libelectrod.Name.t -> Libelectrod.Tuple_set.t -> t -> t

Adds an association to the instance. The name must not be in the instance already.

val mem : Libelectrod.Name.t -> t -> bool

Checks whether a name is already bound in the map.

Accessors

val get_exn : Libelectrod.Name.t -> t -> Libelectrod.Tuple_set.t

May raise Not_found. IT MAY BE BETTER TO RETURN AN EXACT SCOPE OR EVEN A RELATION (TO BE DECIDED WHEN INSTANCES ARE USED IN THE TRANSLATION)

val get : Libelectrod.Name.t -> t -> Libelectrod.Tuple_set.t option

May rather return None. IT MAY BE BETTER TO RETURN AN EXACT SCOPE OR EVEN A RELATION (TO BE DECIDED WHEN INSTANCES ARE USED IN THE TRANSLATION)

val to_list : t -> (Libelectrod.Name.t * Libelectrod.Tuple_set.t) list

Returns the map as an association list. IT MAY BE BETTER TO RETURN AN EXACT SCOPE OR EVEN A RELATION (TO BE DECIDED WHEN INSTANCES ARE USED IN THE TRANSLATION)

val to_map : t -> Libelectrod.Tuple_set.t Libelectrod.Name.Map.t

val rename : (Libelectrod.Atom.t, Libelectrod.Atom.t) Containers.List.Assoc.t -> (Libelectrod.Name.t, Libelectrod.Name.t) Containers.List.Assoc.t -> t -> t

include Libelectrod.Intf.Print.S with type t := t

type t

val pp : t Libelectrod.Fmtc.t

val to_string : t -> string