Module Libelectrod__Name

Names (of relations for instance).

type t

val name : string -> t

Constructors

val dummy : unit -> t

Returns a dummy name. This function is only here for some computations that raise new relations and hence need a temporary, dummy name.

val of_raw_ident : Libelectrod.Raw_ident.t -> t

val univ : t

Reserved name for 'univ' and 'iden'.

val iden : t

val equal : t -> t -> bool

Tells whether two names are the same

val compare : t -> t -> int

val hash : t -> int

val style : Fmt.style

include Libelectrod.Intf.Print.S with type t := t

type t

val pp : t Libelectrod.Fmtc.t

val to_string : t -> string

module Map : CCMap.S with type Map.key = t