Module Libelectrod__Scope

type relation = private

| Plain_relation of Libelectrod.Tuple_set.t * Libelectrod.Tuple_set.t
inv: inf in sup

| Partial_function of int * Libelectrod.Tuple_set.t
int is the domain arity (inv: >= 0); inf = empty

| Total_function of int * Libelectrod.Tuple_set.t
int is the domain arity (inv: >= 0); inf = empty

type t = private

| Exact of Libelectrod.Tuple_set.t
means: lower bound = upper bound

| Inexact of relation

Constructors

val exact : Libelectrod.Tuple_set.t -> t

val plain_relation : Libelectrod.Tuple_set.t -> Libelectrod.Tuple_set.t -> relation

val partial_function : int -> Libelectrod.Tuple_set.t -> relation

val total_function : int -> Libelectrod.Tuple_set.t -> relation

val inexact : relation -> t

val equal : t -> t -> bool

val included_in : Libelectrod.Tuple_set.t -> t -> bool

included_in ts scope tells whether ts is in the scope (meaning it also contains the lower bound of the scope if the latter is inexact.)

val inf : t -> Libelectrod.Tuple_set.t

Return the inf and sup bounds of the scope.

val sup : t -> Libelectrod.Tuple_set.t

val must : t -> Libelectrod.Tuple_set.t

Return the must and may (= sup - inf; computation is cached) bounds of the scope.

val may : t -> Libelectrod.Tuple_set.t

val is_partial : t -> bool

val inferred_arity : t -> int

0 if the arity cannot be inferred (= is unknown), n > 0 otherwise.

val rename : (Libelectrod.Atom.t, Libelectrod.Atom.t) Containers.List.Assoc.t -> t -> t

include Libelectrod.Intf.Print.S with type t := t

type t

val pp : t Libelectrod.Fmtc.t

val to_string : t -> string