Module Libelectrod__Transfo

Specification of transformations.

type ('src, 'dst) t

A transformation conforms to the following signature.

val make : string -> ('src -> 'dst) -> ('src, 'dst) t

Constructor. The first parameter is the name: it must be unique in the whole program.

val name : ('src, 'dst) t -> string

Retrieves the name...

val run : ('src, 'dst) t -> 'src -> 'dst

Runs a transformation

val fby : ('a, 'b) t -> ('b, 'c) t -> ('a, 'c) t

Returns the composition (left-to-right) of two transformations.

val identity : ('a, 'a) t

Identity transformation

type ('src, 'dst) tlist

A list of transformations with given source and target type.

val tlist : ('src, 'dst) t list -> ('src, 'dst) tlist

Creates a list of transformations with a default one .

val get_exn : ('src, 'dst) tlist -> string -> ('src, 'dst) t

Retrieves a transformation. May raise Not_found.